tidy up: splitted material code from core
