+++ demos/demo_grin.py
@@ -30,8 +30,8 @@ import matplotlib
 from distutils.version import StrictVersion
 
 from pyrateoptics.sampling2d import raster
-from pyrateoptics.core.material_grin import IsotropicGrinMaterial
-from pyrateoptics.core.material_isotropic import ConstantIndexGlass
+from pyrateoptics.material.material_grin import IsotropicGrinMaterial
+from pyrateoptics.material.material_isotropic import ConstantIndexGlass
 from pyrateoptics.core import surfShape
 from pyrateoptics.core.optical_element import OpticalElement
 from pyrateoptics.core.surface import Surface
