+++ demos/demo_mirrors.py
@@ -31,8 +31,8 @@ from distutils.version import StrictVersion
 
 
 from pyrateoptics.sampling2d import raster
-from pyrateoptics.core.material_isotropic import ConstantIndexGlass
-from pyrateoptics.core.material_anisotropic import AnisotropicMaterial
+from pyrateoptics.material.material_isotropic import ConstantIndexGlass
+from pyrateoptics.material.material_anisotropic import AnisotropicMaterial
 from pyrateoptics.core import surfShape
 from pyrateoptics.core.optical_element import OpticalElement
 from pyrateoptics.analysis.optical_element_analysis import OpticalElementAnalysis
