+++ demos/demo_zmx.py
@@ -31,7 +31,7 @@ import logging
 logging.basicConfig(level=logging.DEBUG)
 
 from pyrateoptics.core.localcoordinates import LocalCoordinates
-from pyrateoptics.core.material_isotropic import ConstantIndexGlass
+from pyrateoptics.material.material_isotropic import ConstantIndexGlass
 from pyrateoptics.core.globalconstants import standard_wavelength, numerical_tolerance
 from pyrateoptics.core.ray import RayBundle
 from pyrateoptics.io.zmx import ZMXParser
